backend/requirement.txt
fastapi
uvicorn
chromadb
sentence-transformers
google-genai
pydantic
python-dotenv